Move paste utility to API
